Revert "Setup jest"
