+++ .gitignore
@@ -40,4 +40,3 @@ yarn-error.log*
 
 # typescript
 *.tsbuildinfo
-**/firestore-debug.log
