--- firebase.json
@@ -1,10 +1,0 @@
-{
-  "emulators": {
-    "firestore": {
-      "port": 8080
-    },
-    "ui": {
-      "enabled": false
-    }
-  }
-}
