+++ package.json
@@ -9,8 +9,7 @@
     "start": "next start",
     "analyze": "cross-env ANALYZE=true next build",
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
-    "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
-    "test": "yarn firebase emulators:exec --only firestore 'jest --detectOpenHandles --coverage'"
+    "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build"
   },
   "dependencies": {
     "@chakra-ui/icons": "^2.0.19",
@@ -53,7 +52,6 @@
   "devDependencies": {
     "@babel/core": "^7.20.12",
     "@next/bundle-analyzer": "^13.1.1",
-    "@types/jest": "^29.5.3",
     "@types/node": "^18.11.18",
     "@types/nodemailer": "^6.4.7",
     "@types/papaparse": "^5.3.7",
@@ -67,13 +65,10 @@
     "csstype": "^3.1.1",
     "eslint": "^8.31.0",
     "eslint-config-next": "^13.1.6",
-    "firebase-tools": "^12.4.4",
-    "jest": "^29.6.1",
     "postcss": "^8.4.21",
     "prettier": "^2.8.2",
     "prettier-plugin-tailwindcss": "^0.2.3",
     "tailwindcss": "^3.2.4",
-    "ts-jest": "^29.1.1",
     "typescript": "^4.9.4"
   },
   "ct3aMetadata": {
