--- src/server/db/collections/admin/AppliedRoleCollection.ts
@@ -1,9 +1,0 @@
-import { BaseCollection } from './BaseCollection'
-import type { AppliedRole } from '../../models/AppliedRole'
-
-class AppliedRoleCollection extends BaseCollection<AppliedRole> {
-  protected override collectionName = 'applied_roles'
-  protected override objectName = 'applied role'
-}
-
-export const appliedRoleCollection = new AppliedRoleCollection()
