+++ src/server/trpc/router/recruitment/create.ts
@@ -1,8 +1,11 @@
-import { Timestamp } from 'firebase/firestore'
+import { runTransaction, Timestamp } from 'firebase/firestore'
 import { z } from 'zod'
+import { db } from '~/server/db/firebase'
+import { ApplicationStatus } from '~/server/db/models/AppliedRole'
 import logCollection from '~/server/db/collections/LogCollection'
 import { protectedProcedure } from '../../trpc'
-import { CreateManyAppliedRoleController } from '../controllers/recruitment/create_many_applied_roles_controller'
+import appliedRoleCollection from '~/server/db/collections/AppliedRoleCollection'
+import { addUsers } from '../users/createManyUsers'
 
 export const createManyAppliedRoles = protectedProcedure
   .input(
@@ -32,10 +35,38 @@ export const createManyAppliedRoles = protectedProcedure
       ),
     })
   )
-  .mutation(async ({ input: { users, applicants } }) => {
+  .mutation(async ({ input }) => {
     try {
-      const contoller = new CreateManyAppliedRoleController()
-      return await contoller.execute(applicants, users)
+      return await runTransaction(db, async (transaction) => {
+        input.users = input.users.filter(user => user.student_id)
+
+        await addUsers(input.users, transaction)
+        input.applicants.forEach((appliedRole) => {
+          /// Upload first choice
+          const id_one = appliedRoleCollection.generateRandomId()
+          appliedRoleCollection.withTransaction(transaction).set({
+            applicantId: appliedRole.applicantId,
+            id: id_one,
+            rank: 1,
+            role: appliedRole.firstRole,
+            department: appliedRole.firstDepartment,
+            status: ApplicationStatus.PENDING,
+            flag: false,
+          }, id_one)
+
+          /// Upload second choice
+          const id_two = appliedRoleCollection.generateRandomId()
+          appliedRoleCollection.withTransaction(transaction).set({
+            applicantId: appliedRole.applicantId,
+            rank: 2,
+            id: id_two,
+            role: appliedRole.secondRole,
+            department: appliedRole.secondDepartment,
+            status: ApplicationStatus.PENDING,
+            flag: false,
+          }, id_two)
+        })
+      })
     } catch (e) {
       await logCollection.add({
         createdAt: Timestamp.fromDate(new Date()),
