+++ src/server/trpc/router/users/createManyUsers.ts
@@ -1,7 +1,71 @@
 import { protectedProcedure } from '../../trpc'
 import { z } from 'zod'
-import { TRPCError } from '@trpc/server'
-import { CreateManyUserController } from '../controllers/users/create_many_user_controller'
+import { randomUUID } from 'crypto'
+import { sendMultipleEmails } from '../member/helper'
+import { db } from '~/server/db/firebase'
+import logCollection from '~/server/db/collections/LogCollection'
+import { Timestamp, type Transaction, runTransaction } from 'firebase/firestore'
+import userCollection from '~/server/db/collections/UserCollection'
+import { adminAuth } from '~/server/db/admin_firebase'
+
+export async function addUsers(
+  input: {
+    name: string
+    personal_email: string
+    department: string
+    role: string
+    nus_email: string
+    student_id: string
+    resume?: string | undefined
+  }[],
+  transaction: Transaction
+) {
+  /// Step 1: Clean the data by filtering until we have the valid users.
+  const users = await userCollection.queries()
+  const userIds = new Set(users.map((user) => user.id))
+  const userEmails = new Set(users.map((user) => user.email))
+  input = input.filter(
+    (user) =>
+      user.student_id &&
+      !userIds.has(user.student_id) &&
+      !userEmails.has(user.personal_email)
+  )
+
+  /// Step 2: Save the data for all the valid users.
+  const emailData = await Promise.all(
+    input.map(async (user) => {
+      const password = randomUUID().substring(0, 10)
+
+      await adminAuth.createUser({
+        email: user.personal_email,
+        displayName: user.name,
+        uid: user.student_id,
+        password,
+      })
+
+      userCollection.withTransaction(transaction).set(
+        {
+          department: user.department,
+          email: user.nus_email,
+          name: user.name,
+          isAdmin: false,
+          id: user.student_id,
+          role: user.role,
+          resume: user.resume || '',
+          personal_email: user.personal_email,
+        },
+        user.student_id
+      )
+
+      return {
+        email: user.personal_email,
+        password,
+      }
+    })
+  )
+
+  await sendMultipleEmails(emailData)
+}
 
 export const createManyUsers = protectedProcedure
   .input(
@@ -22,12 +86,13 @@ export const createManyUsers = protectedProcedure
   )
   .mutation(async ({ input }) => {
     try {
-      const controller = new CreateManyUserController()
-      return await controller.execute(input)
+      return await runTransaction(db, (transaction) => addUsers(input, transaction))
     } catch (e) {
-      throw new TRPCError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: (e as Error).message
+      await logCollection.add({
+        createdAt: Timestamp.fromDate(new Date()),
+        level: 'WARNING',
+        description: (e as Error).message,
+        title: 'Error while uploading multiple users',
       })
     }
   })
