--- test/setup.ts
@@ -1,9 +1,0 @@
-import { db } from './firebase'
-
-jest.mock('../src/server/db/admin_firebase', () => {
-  return { db, adminAuth: { createUser: jest.fn() } }
-})
-
-jest.mock("../src/server/trpc/router/member/helper", () => ({
-  sendMultipleEmails: jest.fn()
-}))
