--- test/util/dropCollection.ts
@@ -1,6 +1,0 @@
-import { db } from "../firebase";
-
-export async function dropCollection(collectionName: string) {
-    const result = await db.collection(collectionName).get().then((snapshot) => snapshot.docs)
-    await Promise.all(result.map(res => res.ref.delete()))
-}
